requirements.txt
anyio>=3.7.0
click>=8.1.0
fastapi>=0.104.0
h11>=0.14.0
httptools>=0.6.0
idna>=3.4
pydantic>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
sniffio>=1.3.0
starlette>=0.27.0
typing-extensions>=4.8.0
uvicorn>=0.24.0
uvloop>=0.19.0
watchfiles>=0.21.0
websockets>=12.0
openai>=1.3.3
packaging>=23.0
requests>=2.31.0
qdrant-client>=1.7.0
pypdf>=3.17.0
tiktoken>=0.5.0

